requirements.txt
# Core
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.3

# Web scraping
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2

# Image processing
Pillow==10.4.0
imagehash==4.3.1

# Dashboard
streamlit==1.39.0
altair>=5.0.0

# API sources
keepa==1.4.0

# Testing
pytest>=8.0.0
